backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.1.0",
    "aiosqlite>=0.22.1",
    "chromadb>=1.5.0",
    "fastapi>=0.128.2",
    "greenlet>=3.0.0",
    "hdbscan>=0.8.41",
    "httpx>=0.28.1",
    "pillow>=12.1.0",
    "pynndescent>=0.6.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "sentencepiece>=0.1.99",
    "sqlalchemy>=2.0.46",
    "torch>=2.5.0",
    "transformers==4.46.0",
    "umap-learn>=0.5.11",
    "uvicorn>=0.27.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
